Move tableProps factory function out of component
